pyproject.toml
[project]
name = "causal-armor"
version = "0.1.0"
description = "Efficient Indirect Prompt Injection guardrails via causal attribution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Prashant Kulkarni", email = "kulkarniprashants@gmail.com"},
]
keywords = [
    "llm",
    "security",
    "prompt-injection",
    "causal-attribution",
    "guardrails",
    "ai-safety",
    "tool-use",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.25",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
gemini = ["google-genai>=1.0"]
litellm = ["litellm>=1.0"]
all = ["openai>=1.0", "anthropic>=0.30", "google-genai>=1.0", "litellm>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.8",
    "mypy>=1.13",
    "build>=1.0",
    "twine>=6.0",
]

[project.urls]
Homepage = "https://github.com/prashantkul/causal-armor"
Repository = "https://github.com/prashantkul/causal-armor"
Documentation = "https://github.com/prashantkul/causal-armor/tree/master/docs"
Issues = "https://github.com/prashantkul/causal-armor/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["causal_armor"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "openai",
    "openai.*",
    "anthropic",
    "anthropic.*",
    "google",
    "google.*",
    "google.genai",
    "google.genai.*",
    "litellm",
    "litellm.*",
    "respx",
    "respx.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "causal_armor.providers.gemini"
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
